requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.1.0
Pillow==10.1.0
openai==1.3.0
google-generativeai==0.3.2
transformers==4.35.2
torch==2.1.1
torchaudio==2.1.1
opencv-python==4.8.1.78
face-recognition==1.3.0
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
whisper==1.1.10
pydub==0.25.1
